requirements.txt
langchain==0.3.27
langchain-openai==0.3.33
langchain-chroma==0.2.6
langchain-text-splitters==0.3.9
chromadb==1.0.20
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.41.1